[#13] word hwp file diff

+++ src/main/java/Coop/controller/FileController.java
@@ -27,6 +27,7 @@ import Coop.mapper.NoticeUserMapper;
 import Coop.mapper.PngFilesMapper;
 import Coop.mapper.ProUserMapper;
 import Coop.mapper.ProjectMapper;
+import Coop.mapper.TextDiffMapper;
 import Coop.mapper.UserMapper;
 import Coop.model.Comment;
 import Coop.model.File;
@@ -37,9 +38,11 @@ import Coop.model.NoticeUser;
 import Coop.model.PngFiles;
 import Coop.model.Pro_User;
 import Coop.model.Project;
+import Coop.model.TextDiff;
 import Coop.model.User;
 import Coop.model.UserKey;
 import Coop.service.FileService;
+import Coop.service.FileToText;
 import Coop.service.IOSPushService;
 import Coop.service.ImageCompareService;
 import Coop.service.MobileAuthenticationService;
@@ -70,6 +73,8 @@ public class FileController {
 	@Autowired NoticeUserMapper noticeUserMapper;
 	@Autowired MobileAuthenticationService mobileAuthenticationService;
 	@Autowired IOSPushService iosPushService;
+	@Autowired FileToText fileToText;
+	@Autowired TextDiffMapper textDiffMapper;
 	
 	
 	@RequestMapping(value = "/{projectId}/{userId}/create.do",method = RequestMethod.GET)
@@ -93,10 +98,21 @@ public class FileController {
 	@RequestMapping(value = "/{fileId}/{proId}/detail.do",method = RequestMethod.GET)
 	public String detail(@PathVariable String fileId,@PathVariable String proId,Model model) {
 		
+		
 		model.addAttribute("project", projectMapper.selectByProjectId(Integer.parseInt(proId)));
 		model.addAttribute("file",fileMapper.selectById(Integer.parseInt(fileId)));
 		model.addAttribute("commentList",commentMapper.selectByFileId(Integer.parseInt(fileId)));
 		model.addAttribute("pngs",pngFilesMapper.selectByFileId(Integer.parseInt(fileId)));
+		List<TextDiff> tf = textDiffMapper.selectByfileId(Integer.parseInt(fileId));
+		if(tf.size()>0){
+			//System.out.println(tf.get(0).getContent());
+			String[] showDiff = tf.get(0).getContent().split("\n");
+			for(int i=0;i<showDiff.length;i++){
+				//System.out.println(showDiff[i]);
+			}
+			model.addAttribute("showDiff", showDiff);
+		}
+		
         return "layout/file/detail";
     }
 	@RequestMapping(value = "{noticeId}/{fileId}/{proId}/detailCheck.do",method = RequestMethod.GET)
@@ -115,6 +131,11 @@ public class FileController {
 		model.addAttribute("file",fileMapper.selectById(Integer.parseInt(fileId)));
 		model.addAttribute("commentList",commentMapper.selectByFileId(Integer.parseInt(fileId)));
 		model.addAttribute("pngs",pngFilesMapper.selectByFileId(Integer.parseInt(fileId)));
+		List<TextDiff> tf = textDiffMapper.selectByfileId(Integer.parseInt(fileId));
+		if(tf.size()>0){
+			String[] showDiff = tf.get(0).getContent().split("\n");
+			model.addAttribute("showDiff", showDiff);
+		}
         return "layout/file/detail";
     }
 	
@@ -145,17 +166,60 @@ public class FileController {
 			 java.io.File dir = new java.io.File(path);
 			 if (!dir.exists()) { //폴더 없으면 폴더 생성
 	            dir.mkdirs();
-	         }
+	         } 
+			 
 			 fileService.writeFile(uploadedFile,"C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName()+"/",file.getFileName());
-			 FilePngCount filePngCount = new FilePngCount();
-			 filePngCount.setFileId(file.getId());
-			 if(file.getFileName().endsWith(".pptx")){
-				 filePngCount.setPngCount(pptChange.convert("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName(), "C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName()+"/"+file.getFileName(), file.getFileName()));
+			 if(!(file.getFileName().endsWith(".pptx")||file.getFileName().endsWith(".pdf"))){
+				 
+				    notice.setProjectId(Integer.parseInt(projectId));
+					notice.setUserId(userService.getCurrentUser().getId());
+					String noDes = userService.getCurrentUser().getName()+"님이 "+"새로운 파일을 업로드 했습니다";
+					notice.setDes(noDes);
+					notice.setFileId(file.getId());
+					noticeMapper.insert(notice);
+					
+					List<User> proUser = proUserMapper.selectByProjectId(Integer.parseInt(projectId));
+					
+					for(int i=0;i<proUser.size();i++){
+						if(!userService.getCurrentUser().getId().equals(proUser.get(i).getId())){
+							NoticeUser noticeUser = new NoticeUser();
+							noticeUser.setId(notice.getId());
+							noticeUser.setProjectId(Integer.parseInt(projectId));
+							noticeUser.setMember(proUser.get(i).getId());
+							
+							noticeUserMapper.insert(noticeUser);
+							UserKey userKey = userMapper.selectByKey(proUser.get(i).getId());
+							if(userKey!=null){
+								iosPushService.push(userKey.getUserkey(),userService.getCurrentUser().getName()+" 님이 "+proj.getName()+"에 새로운 파일을 업로드 했습니다.");
+							}
+						}
+						
+								
+					}
+					NoticeUser noticeUser=  new NoticeUser();
+					noticeUser.setProjectId(Integer.parseInt(projectId));
+					noticeUser.setMember(userService.getCurrentUser().getId());
+					model.addAttribute("noticeList",noticeMapper.select(noticeUser));
+					model.addAttribute("project",projectMapper.selectByProjectId(Integer.parseInt(projectId)));
+					model.addAttribute("fileList", fileMapper.selectByProjectId(Integer.parseInt(projectId)));
+					model.addAttribute("userList",userMapper.selectProject(Integer.parseInt(projectId)));
+					model.addAttribute("pro_user",proUserMapper.selectByProjectId(Integer.parseInt(projectId)));
+			        return "layout/project/info";
 			 }
 			 else{
-				 filePngCount.setPngCount(pdfChange.changePDF("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName()+"/"+file.getFileName()));
-				 filePngCountMapper.insert(filePngCount);
+				 
+				 FilePngCount filePngCount = new FilePngCount();
+				 filePngCount.setFileId(file.getId());
+				 if(file.getFileName().endsWith(".pptx")){
+					 filePngCount.setPngCount(pptChange.convert("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName(), "C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName()+"/"+file.getFileName(), file.getFileName()));
+				 }
+				 else{
+					 filePngCount.setPngCount(pdfChange.changePDF("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName()+"/"+file.getFileName()));
+					 filePngCountMapper.insert(filePngCount);
+				 }
 			 }
+			 
+			 
 			 
 		}
 		notice.setProjectId(Integer.parseInt(projectId));
@@ -198,6 +262,8 @@ public class FileController {
 	public String create2(@PathVariable String projectId,@PathVariable String userId,@RequestParam("des") String des,
 			@RequestParam("file") MultipartFile uploadedFile,Model model,@PathVariable String fileId) throws IOException, CommunicationException, KeystoreException, InvalidDeviceTokenFormatException {
 		List<FileInner> fileInner = fileInnerMapper.selectByRefFileId(Integer.parseInt(fileId));
+	
+		
 		
 		Notice notice = new Notice();
 		if(fileInner.size()<1){
@@ -227,10 +293,28 @@ public class FileController {
 				 file.setDes(des);
 				 file.setRefFile(Integer.parseInt(fileId));
 				 fileInnerMapper.insert(file);
-				 pngFilesMapper.delete(Integer.parseInt(fileId));
+				
 				 //fileService.writeFile(uploadedFile,"C:\\Users\\USER\\FileSave\\"+resFile.getFileName()+"\\",file.getFileName());
 				 fileService.writeFile(uploadedFile,"C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"\\",file.getFileName());
 				 //pngCount = pdfChange.changePDF("C:/Users/USER/FileSave/"+resFile.getFileName()+"/"+file.getFileName());
+				 if(!(file.getFileName().endsWith(".pptx") || file.getFileName().endsWith(".pdf"))){
+					 String[] diffText = fileToText.compareText("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"/"+resFile.getFileName(), "C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"/"+file.getFileName());
+					 String diff = fileService.compareText(diffText[0], diffText[1]);
+					 TextDiff td = new TextDiff();
+					 td.setFileId(resFile.getId());
+					 td.setContent(diff);
+					 td.setDes(des);
+					 td.setUserId(userId);
+					 td.setUserName(userService.getCurrentUser().getName());
+					 textDiffMapper.insert(td);
+					 model.addAttribute("project",projectMapper.selectByProjectId(Integer.parseInt(project)));
+					 model.addAttribute("fileList",fileMapper.selectByProjectId(Integer.parseInt(project)));
+					
+					 
+					 return "layout/project/info";
+					 
+				 }
+				 pngFilesMapper.delete(Integer.parseInt(fileId));
 				 if(file.getFileName().endsWith(".pptx")){
 					 pngCount = pptChange.convert("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName(), "C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"/"+file.getFileName(), file.getFileName());
 				 }else{
@@ -241,9 +325,9 @@ public class FileController {
 				 //imageCompareService.compare(resFile.getFileName(), file.getFileName(), pngCount, path);
 				 //result = imageCompareService.compare("skh", file.getFileName(),1, path+"/");
 				 //result = imageCompareService.compare(resFile.getFileName().substring(0,resFile.getFileName().indexOf('.')), file.getFileName(),2, path+"/");
-				 result = imageCompareService.compare2(resFile.getFileName(),file.getFileName(),2, path+"/");
+				 result = imageCompareService.compare2(resFile.getFileName(),file.getFileName(),pngCount, path+"/");
 			}
-			PngFiles[] png = new PngFiles[result.length];
+			PngFiles[] png = new PngFiles[result.length-1];
 			System.out.println(png.length);
 			for(int i=0;i<png.length;i++){
 				png[i] = new PngFiles();
@@ -282,6 +366,53 @@ public class FileController {
 				 file.setDes(des);
 				 file.setRefFile(Integer.parseInt(fileId));
 				 fileInnerMapper.insert(file);
+				 fileService.writeFile(uploadedFile,"C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"\\",file.getFileName());
+				 if(!(file.getFileName().endsWith(".pptx") || file.getFileName().endsWith(".pdf"))){
+					 String[] diffText = fileToText.compareText("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"/"+resFile.getFileName(), "C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"/"+file.getFileName());
+					 String diff = fileService.compareText(diffText[0], diffText[1]);
+					 TextDiff td = new TextDiff();
+					 td.setFileId(resFile.getId());
+					 td.setContent(diff);
+					 td.setDes(des);
+					 td.setUserId(userId);
+					 td.setUserName(userService.getCurrentUser().getName());
+					 textDiffMapper.insert(td);
+					    notice.setProjectId(Integer.parseInt(projectId));
+						notice.setUserId(userService.getCurrentUser().getId());
+						String noDes = userService.getCurrentUser().getName()+"님이 "+"새로운 파일을 업로드 했습니다";
+						notice.setDes(noDes);
+						notice.setFileId(resFile.getId());
+						noticeMapper.insert(notice);
+						List<User> proUser = proUserMapper.selectByProjectId(Integer.parseInt(projectId));
+						
+						for(int i=0;i<proUser.size();i++){
+							if(!userService.getCurrentUser().getId().equals(proUser.get(i).getId())){
+								NoticeUser noticeUser = new NoticeUser();
+								noticeUser.setId(notice.getId());
+								noticeUser.setProjectId(Integer.parseInt(projectId));
+								noticeUser.setMember(proUser.get(i).getId());
+								
+								noticeUserMapper.insert(noticeUser);
+								UserKey userKey = userMapper.selectByKey(proUser.get(i).getId());
+								if(userKey!=null){
+									iosPushService.push(userKey.getUserkey(),userService.getCurrentUser().getName()+" 님이 "+resFile.getFileName()+"에 새 버전 파일을 업로드 했습니다");
+								}
+							}
+							
+									
+						}
+						NoticeUser noticeUser=  new NoticeUser();
+						noticeUser.setProjectId(Integer.parseInt(projectId));
+						noticeUser.setMember(userService.getCurrentUser().getId());
+						model.addAttribute("noticeList",noticeMapper.select(noticeUser));
+						
+						model.addAttribute("project",projectMapper.selectByProjectId(Integer.parseInt(projectId)));
+						model.addAttribute("fileList", fileMapper.selectByProjectId(Integer.parseInt(projectId)));
+						model.addAttribute("userList",userMapper.selectProject(Integer.parseInt(projectId)));
+						model.addAttribute("pro_user",proUserMapper.selectByProjectId(Integer.parseInt(projectId)));
+				        return "layout/project/info";
+					 
+				 }
 				 pngFilesMapper.delete(Integer.parseInt(fileId));
 				 //fileService.writeFile(uploadedFile,"C:\\Users\\USER\\FileSave\\"+resFile.getFileName()+"\\",file.getFileName());
 				 fileService.writeFile(uploadedFile,"C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+resFile.getFileName()+"\\",file.getFileName());
@@ -415,6 +546,11 @@ public class FileController {
 		model.addAttribute("file",fileMapper.selectById(Integer.parseInt(fileId)));
 		model.addAttribute("commentList",commentMapper.selectByFileId(Integer.parseInt(fileId)));
 		model.addAttribute("pngs",pngFilesMapper.selectByFileId(Integer.parseInt(fileId)));
+		List<TextDiff> tf = textDiffMapper.selectByfileId(Integer.parseInt(fileId));
+		if(tf.size()>0){
+			String[] showDiff = tf.get(0).getContent().split("\n");
+			model.addAttribute("showDiff", showDiff);
+		}
         return "layout/file/detail";
 		
         
@@ -429,6 +565,7 @@ public class FileController {
 			  fileInnerMapper.delete(file.getId());
 			  fileMapper.delete(Integer.parseInt(fileId));
 			  fileService.deleteFolder("C:/Users/USER/Documents/website/neonWork/Coop/src/main/webapp/res/FileSave/"+file.getFileName());
+			  textDiffMapper.delete(Integer.parseInt(fileId));
 			  	NoticeUser noticeUser=  new NoticeUser();
 				noticeUser.setProjectId(Integer.parseInt(projectId));
 				noticeUser.setMember(userService.getCurrentUser().getId());
@@ -466,14 +603,24 @@ public class FileController {
 				model.addAttribute("file",fileMapper.selectById(Integer.parseInt(fileId)));
 				model.addAttribute("commentList",commentMapper.selectByFileId(Integer.parseInt(fileId)));
 				model.addAttribute("pngs",pngFilesMapper.selectByFileId(Integer.parseInt(fileId)));
-		        return "layout/file/detail";
+				List<TextDiff> tf = textDiffMapper.selectByfileId(Integer.parseInt(fileId));
+				if(tf.size()>0){
+					String[] showDiff = tf.get(0).getContent().split("\n");
+					model.addAttribute("showDiff", showDiff);
+				}
+				return "layout/file/detail";
 		  }
 		  else{
 			    model.addAttribute("project", projectMapper.selectByProjectId(Integer.parseInt(projectId)));
 				model.addAttribute("file",fileMapper.selectById(Integer.parseInt(fileId)));
 				model.addAttribute("commentList",commentMapper.selectByFileId(Integer.parseInt(fileId)));
 				model.addAttribute("pngs",pngFilesMapper.selectByFileId(Integer.parseInt(fileId)));
-		        return "layout/file/detail";
+				List<TextDiff> tf = textDiffMapper.selectByfileId(Integer.parseInt(fileId));
+				if(tf.size()>0){
+					String[] showDiff = tf.get(0).getContent().split("\n");
+					model.addAttribute("showDiff", showDiff);
+				}
+				return "layout/file/detail";
 		  }
 		
         
@@ -572,6 +719,15 @@ public class FileController {
 		if(mobileAuthenticationService.AuthenticationUser(userService.getCurrentUser())){
 			model.addAttribute("file", fileMapper.selectById(Integer.parseInt(fileId)));
 			model.addAttribute("pngs",pngFilesMapper.selectByFileId(Integer.parseInt(fileId)));
+			List<TextDiff> tf = textDiffMapper.selectByfileId(Integer.parseInt(fileId));
+			if(tf.size()>0){
+				//System.out.println(tf.get(0).getContent());
+				String[] showDiff = tf.get(0).getContent().split("\n");
+				for(int i=0;i<showDiff.length;i++){
+					//System.out.println(showDiff[i]);
+				}
+				model.addAttribute("showDiff", showDiff);
+			}
 	        return "webview/pngview";
 		}
 		else{
